rules: fix type of sid
In postgresql the integer are signed and integer field is a u32.
As a result, a signature with SID above 2^31 can not be imported
but Suricata supports signature as a 32 bit unsigned we can fail
importing valid rules.

+++ rules/models.py
@@ -2483,7 +2483,7 @@ class RangeCheckIntegerFields(models.Model):
 
 class Rule(RangeCheckIntegerFields, Transformable, Cache):
     GROUP_BY_CHOICES = (('by_src', 'by_src'), ('by_dst', 'by_dst'))
-    sid = models.IntegerField(primary_key=True)
+    sid = models.BigIntegerField(primary_key=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     msg = models.CharField(max_length=1000)
     group = models.BooleanField(default=False)
